clean up cookies controller/route structure, rename cookies-banner to cookies
