--- src/ui/public/js/cookiesBanner.js
@@ -1,2 +1,0 @@
-var DTFS;!function(){var e=document.querySelector(".js-hide"),t=document.querySelector(".govuk-cookie-banner"),o=document.querySelector(".js-cookies-button-accept"),n=document.querySelector(".js-cookies-button-reject"),c=document.querySelector(".change-cookies-submit-button"),i=function(){const e=new Date;return e.setDate(e.getDate()+1),e.toGMTString(),e},u=function(){document.cookie="optionalCookies=false; SameSite=Lax; expires="+i()},r=function(){document.cookie="optionalCookies=true; SameSite=Lax; expires="+i()};n&&n.addEventListener("click",(function(e){u()})),o&&o.addEventListener("click",(function(e){r()})),document.cookie.includes("optionalCookies")||(t.removeAttribute("hidden"),t.setAttribute("tabindex","-1"),t.focus(),t.addEventListener("blur",(function(){t.removeAttribute("tabindex")}))),e&&e.addEventListener("click",(function(e){t.setAttribute("hidden","hidden"),e.preventDefault()})),c&&c.addEventListener("click",(function(e){if(document.querySelector('input[name="optionalCookies"]:checked')){const e=document.querySelector('input[name="optionalCookies"]:checked').value;"reject"===e&&(console.log("======== rejecting"),u()),"accept"===e&&(console.log("======== accepting"),r())}})),(DTFS=void 0===DTFS?{}:DTFS).cookiesBanner={}}();
-//# sourceMappingURL=cookiesBanner.js.map
