+++ src/ui/server/controllers/root/cookies-consent/index.test.ts
@@ -1,6 +1,6 @@
 import post from '.';
-import { mockReq, mockRes } from '../../test-mocks';
-import { Request, Response } from '../../../types';
+import { mockReq, mockRes } from '../../../test-mocks';
+import { Request, Response } from '../../../../types';
 
 describe('controllers/cookies-consentt', () => {
   let req: Request;
