--- src/ui/server/routes/cookies-consent/index.test.ts
@@ -1,19 +1,0 @@
-import { post } from '../../test-mocks/mock-router';
-import { ROUTES } from '../../constants';
-import cookiesConsentPost from '../../controllers/cookies-consent';
-
-describe('routes/cookies-consent', () => {
-  beforeEach(() => {
-    require('.'); // eslint-disable-line global-require
-  });
-
-  afterEach(() => {
-    jest.resetAllMocks();
-  });
-
-  it('should setup all routes', () => {
-    expect(post).toHaveBeenCalledTimes(1);
-
-    expect(post).toHaveBeenCalledWith(ROUTES.COOKIES_CONSENT, cookiesConsentPost);
-  });
-});
