+++ src/ui/server/routes/index.ts
@@ -1,9 +1,7 @@
 import root from './root';
-import cookiesConsent from './cookies-consent';
 import quote from './quote';
 import application from './application';
 
 export const rootRoute = root;
-export const cookiesConsentRoute = cookiesConsent;
 export const quoteRoutes = quote;
 export const applicationRoutes = application;
