+++ src/ui/server/routes/root/index.test.ts
@@ -1,7 +1,8 @@
 import { get, post } from '../../test-mocks/mock-router';
 import { ROUTES } from '../../constants';
 import rootGet from '../../controllers/root';
-import { get as cookiesGet, post as cookiesPost } from '../../controllers/quote/cookies';
+import { get as cookiesGet, post as cookiesPost } from '../../controllers/root/cookies';
+import cookiesConsentPost from '../../controllers/root/cookies-consent';
 import problemWithServiceGet from '../../controllers/quote/problem-with-service';
 
 describe('routes/index', () => {
@@ -15,12 +16,14 @@ describe('routes/index', () => {
 
   it('should setup all routes', () => {
     expect(get).toHaveBeenCalledTimes(3);
-    expect(post).toHaveBeenCalledTimes(1);
+    expect(post).toHaveBeenCalledTimes(2);
 
     expect(get).toHaveBeenCalledWith(ROUTES.ROOT, rootGet);
 
     expect(get).toHaveBeenCalledWith(ROUTES.COOKIES, cookiesGet);
     expect(post).toHaveBeenCalledWith(ROUTES.COOKIES, cookiesPost);
+
+    expect(post).toHaveBeenCalledWith(ROUTES.COOKIES_CONSENT, cookiesConsentPost);
 
     expect(get).toHaveBeenCalledWith(ROUTES.PROBLEM_WITH_SERVICE, problemWithServiceGet);
   });
