+++ src/ui/templates/index.njk
@@ -44,7 +44,7 @@
     {% include "partials/footer.njk" %}
 
     <script src="/assets/js/govukFrontend.js" type="text/javascript"></script>
-    <script src="/assets/js/cookiesBanner.js" type="text/javascript"></script>
+    <script src="/assets/js/cookies.js" type="text/javascript"></script>
 
     {% if cookieConsent %}
       <script async src="https://www.googletagmanager.com/gtag/js?id={{ googleAnalyticsId }}"></script>
